feat: move mapAtom to connectionTree module
