+++ frontend/src/utils/index.ts
@@ -17,7 +17,6 @@ export * from "./tab";
 export * from "./task";
 export * from "./principal";
 export * from "./policy";
-export * from "./sql-editor";
 export * from "./sheet";
 export * from "./pev2";
 export * from "./web-storage";
